- Reading from bundle details API - fixing images aspect ratio

+++ app/src/main/java/com/tedmob/africell/data/api/dto/BundleInfo.kt
@@ -10,7 +10,7 @@ data class BundleInfo(
     @field:[Expose SerializedName("activate")]
     val activate: Boolean?,
     @field:[Expose SerializedName("bundleId")]
-    val bundleId: Int?,
+    val bundleId: Long?,
     @field:[Expose SerializedName("category")]
     val category: String?,
     @field:[Expose SerializedName("commercialName")]
